devops/08-observabilidade/requirements.txt
# Dependências do Módulo 8 — Observabilidade
# Instalar com: pip install -r requirements.txt

# API e servidor web
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2

# Métricas
prometheus-client==0.21.0

# OpenTelemetry (tracing, métricas, logs)
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0
opentelemetry-instrumentation-httpx==0.49b0
opentelemetry-instrumentation-logging==0.49b0
opentelemetry-exporter-otlp-proto-grpc==1.28.0

# Logging estruturado
structlog==24.4.0

# Testes e CLI de apoio
pytest==8.3.3
pytest-asyncio==0.24.0
rich==13.9.3
pyyaml==6.0.2
requests==2.32.3
